Add files via upload
a) Create multiple robots.
b) Associate each robot with a unique number.
a) Move a selected robot in terrain based on the input. If there is already a robot at the destination cell, the robot should stop in the previous cell.
b) Display the current location of the selected robot in the terrain.

The Robo.py hold the robo.py  adding robots, moving them, and getting their current positions.
The test.py hold all the test cases required for the code.